requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Database
sqlmodel>=0.0.8
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Market Data
polygon-api-client==1.13.3
pandas>=1.3.0
python-dotenv>=0.19.0

# Utilities
pydantic==2.6.1
pydantic-settings==2.2.1
httpx==0.27.0
python-dateutil==2.9.0.post0

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
